+++ solys2tracker/tabs.py
@@ -31,7 +31,7 @@ except:
     from common import add_spacer
     from bodywidgets import BodyMenuWidget, BodyTrackWidget, BodyCrossWidget, BodyBlackWidget
     from configwidgets import ConnectionWidget, ConfigNavBarWidget, SpiceWidget, LogWidget, \
-        AdjustWidget
+        AdjustWidget, PositionWidget
 
 """___Authorship___"""
 __author__ = 'Javier Gatón Herguedas'
@@ -141,6 +141,7 @@ class ConfigurationWidget(QtWidgets.QWidget):
         SPICE = 1
         LOG = 2
         ADJUST = 3
+        MOVE_POS = 4
 
     def _change_tab(self, page: PageEnum):
         """
@@ -159,8 +160,10 @@ class ConfigurationWidget(QtWidgets.QWidget):
             self.content_w = SpiceWidget(self, self.session_status)
         elif page == ConfigurationWidget.PageEnum.LOG:
             self.content_w = LogWidget(self, self.session_status)
-        else:
+        elif page == ConfigurationWidget.PageEnum.ADJUST:
             self.content_w = AdjustWidget(self, self.session_status)
+        else:
+            self.content_w = PositionWidget(self, self.session_status)
         self.main_layout.addWidget(self.content_w, 1)
 
     def change_tab_connection(self) -> None:
@@ -186,6 +189,12 @@ class ConfigurationWidget(QtWidgets.QWidget):
         Change the tab to the ADJUST tab.
         """
         self._change_tab(ConfigurationWidget.PageEnum.ADJUST)
+
+    def change_tab_move_pos(self) -> None:
+        """
+        Change the tab to the MOVE_POS tab.
+        """
+        self._change_tab(ConfigurationWidget.PageEnum.MOVE_POS)
 
 class SunTabWidget(QtWidgets.QWidget, ifaces.IBodyTabWidget, metaclass=noconflict.makecls()):
     """