+++ utils/solys2sim.py
@@ -32,6 +32,7 @@ def server_thread(conn: socket.socket):
         data = conn.recv(1024)
         if data:
             empties = 0
+            time.sleep(0.1)
             print(data)
             cmd = str(data)[2:4]
             if cmd == "TI":
@@ -39,11 +40,14 @@ def server_thread(conn: socket.socket):
             elif cmd == "PO":
                 last_po_time = time.time()
                 vals = str(data)[2:-3].split()
-                if int(vals[1]) == 0:
-                    current_azimuth = float(vals[2])
+                if len(vals) == 1:
+                    ret = "PO {} {}".format(current_azimuth, current_zenith)
                 else:
-                    current_zenith = float(vals[2])
-                ret = "PO"
+                    if int(vals[1]) == 0:
+                        current_azimuth = float(vals[2])
+                    else:
+                        current_zenith = float(vals[2])
+                    ret = "PO"
             elif cmd == "CP":
                 current_po_time = time.time()
                 if last_po_time == None or last_po_time + DELAY <= current_po_time: