+++ NEWS.md
@@ -2,3 +2,4 @@
 
 - Logfile improved. New file for each process, with UTC datetime in name.
 - Model ConnectionStatus refactorized to SessionStatus
+- Send Position page in configuration tab.